Corrigido bug na exclusão de agendamentos

+++ src/main/java/br/com/dio/barbershopui/service/impl/ScheduleService.java
@@ -5,7 +5,11 @@ import br.com.dio.barbershopui.repository.IScheduleRepository;
 import br.com.dio.barbershopui.service.IScheduleService;
 import br.com.dio.barbershopui.service.query.IScheduleQueryService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -16,15 +20,17 @@ public class ScheduleService implements IScheduleService {
 
     @Override
     public ScheduleEntity save(final ScheduleEntity entity) {
-        queryService.verifyIfScheduleExists(entity.getStartAt(),entity.getEndAt());
+        queryService.verifyIfScheduleExists(entity.getStartAt(), entity.getEndAt());
 
         return repository.save(entity);
     }
 
     @Override
     public void delete(long id) {
-            queryService.findById(id);
-            repository.deleteById(id);
-
+        Optional<ScheduleEntity> schedule = repository.findById(id);
+        if (!schedule.isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Agendamento não encontrado");
+        }
+        repository.deleteById(id);
     }
 }
